Add Volume ACL Permissions: DB
ACL implementation brings a mechanism of access permission management into
Cinder project. Currently there is only one type of volume access permission,
which is full aceess(The user can do any operations on the volume.).
If a volume is created by a user, it grants all access of this user and
the users with super authoritzations, like administrators.
With ACL implementation, it is possible for different users to have different
levels of volume access permissions. ACL can differentiate users in terms of
granting different access permissions to volumes, which can establish
a foudation for sharing volumes, transferring volumes,
attaching multiple VM to a volume, etc.

Access permissions:

The owner of the volume and the administrators always have the full
access.
The owner, the administrators and the users with Read permission --
can read the permission of the volume.
The owner, the administrators and  the users with Write permission --
can edit the permission of the volume.
The owner and the administrators -- can give the volume access permission
to a single user, a group of users, all users.
The user can check his own access permission to a specific volume.
The permissions of the administrators are unable to edit by any other user.
The permission of the creator can be modified by administrators only.

Access permission can be classifies as the following:

Read: the user can only read the volume, including reading the content
	of the volume, the volume properties and the volume metadata.
	Indicated as 1.
Write: the user can read and write the volume. The user can do any operations
	on this volume, including all the API operations.
	This permission includes Read permission.
	Indicated as 2.
Permission Read: the user can read the permission list for this volume and
	be able to see who can access this volume with which permission level.
	Indicated as 3.
Permission Write: the user can read and edit the permission list for
	this volume and be able to see and edit who can access this volume
	with which permission level. This permission includes Permission Read.
	Indicated as 4.
Read+Permission Read: Indicated as 5.
Write+Permission Read: Indicated as 6.
Full access: This permission inludes all the other permissions and
	equals to Write+Permission Write.
	Indicated as 7.
Read+Permission Write: Indicated as 8.

Implementation:

Part 1. DB
    Add a new table called VolumeACLPermission with columns:
    id - the primary key of the row;
    volume_id - indicates volume's ID;
    entity_id - indicates a user or group ID which the permission
    affects or 'everyone' keyword for all users;
    type - "user" means entity_id is user_id; "group" means
    entity_id is group_id;
    access_permission - indicates access permission level. (from 1 to 7 as
    described in Access Permission above)

    Use cases:
    * The owner(user_id1) grants Write access to user_id2. The
    corresponding record:
        {id: 1, volume_id: 'volume_id1', entity_id: 'user_id2', type:
        'user', access_permission: 2}
    * The owner(user_id1) grants Read access to all users(everyone). The
    corresponding record:
        {id: 2, volume_id: 'volume_id1', entity_id: 'everyone', type:
        'user', access_permission: 1}
    * The owner(user_id1) grants Read+Permission Read access to all the users
    in a group(group_id1). The corresponding record:
        {id: 3, volume_id: 'volume_id1', entity_id: 'group_id1', type:
        'group', access_permission: 5}

blueprint volume-acl

Change-Id: I5e64247ff38c060a2278b3718205834d306f5028