Add Volume ACL Permissions: Volume ACL Module
ACL implementation brings a mechanism of access permission management into
Cinder project. Currently there is only one type of volume access permission,
which is full aceess(The user can do any operations on the volume.).
If a volume is created by a user, it grants all access of this user and
the users with super authoritzations, like administrators.
With ACL implementation, it is possible for different users to have different
levels of volume access permissions. ACL can differentiate users in terms of
granting different access permissions to volumes, which can establish
a foudation for sharing volumes, transferring volumes,
attaching multiple VM to a volume, etc.

Access permissions:

The owner of the volume and the administrators always have the full
access.
The owner, the administrators and the users with Read permission --
can read the permission of the volume.
The owner, the administrators and  the users with Write permission --
can edit the permission of the volume.
The owner and the administrators -- can give the volume access permission
to a single user, a group of users, all users.
The user can check his own access permission to a specific volume.
The permissions of the administrators are unable to edit by any other user.
The permission of the creator can be modified by administrators only.

Access permission can be classifies as the following:

Read: the user can only read the volume, including reading the content
	of the volume, the volume properties and the volume metadata.
	Indicated as 1.
Write: the user can read and write the volume. The user can do any operations
	on this volume, including all the API operations.
	This permission includes Read permission.
	Indicated as 2.
Permission Read: the user can read the permission list for this volume and
	be able to see who can access this volume with which permission level.
	Indicated as 3.
Permission Write: the user can read and edit the permission list for
	this volume and be able to see and edit who can access this volume
	with which permission level. This permission includes Permission Read.
	Indicated as 4.
Read+Permission Read: Indicated as 5.
Write+Permission Read: Indicated as 6.
Full access: This permission inludes all the other permissions and
	equals to Write+Permission Write.
	Indicated as 7.
Read+Permission Write: Indicated as 8.

Implementation:

Part 2. Volume ACL Module
    Add CRUD for VolumeACLPermission.
    Add an interface to work with users and user groups data.
    Not all the openstack deployments use Keystone as Identity Service.
    So it is used here by default, but you can implement your own class depending
    on another Identity Service and specify it in the config.
    Implemented class should contains the following methods:
        * check_user_in_group(self, user_id, group_id);
        * check_user_is_admin(self, user_id, projects)
            checks if user is admin of any of projects in projects list;
        * get_user(self, subject)
            returns existing(means validated) user ID or 'everyone' key word,
            where subject can be user ID or name, or fails with NotFound
            exception;
        * get_group(self, subject)
            returns existing(means validated) group ID or 'everyone' key word,
            where subject is group ID, or fails with NotFound exception.
    Also you can use another implemented class, but there will be
    no functionality with user groups and admins.

    Note, that in a class for work through Keystone Identity Service
    v2.0 keystone client is used as well as v3 keystone client.
    It's due to some functions are implemented in the only version of
    keystone client, e.g. users.check_in_group is implemented in v3
    version only and then we should use v3 keystone client, but
    there is no roles.roles_for_user or any analogs in v3 version,
    in v2.0 only, then we should use v2.0 version for checking
    user(other user, not user_id from context) is admin.
    This will be fixed when it's possible.

blueprint volume-acl

Change-Id: I1537e1462ddb08f72b32b96a085e3ed98efaad48