Add Volume ACL Permissions: API extension
ACL implementation brings a mechanism of access permission management into
Cinder project. Currently there is only one type of volume access permission,
which is full access(The user can do any operations on the volume.).
If a volume is created by a user, it grants all access of this user and
the users with super authoritzations, like administrators.
With ACL implementation, it is possible for different users to have different
levels of volume access permissions. ACL can differentiate users in terms of
granting different access permissions to volumes, which can establish
a foudation for sharing volumes, transferring volumes,
attaching multiple VM to a volume, etc.

Access permissions:

The owner of the volume and the administrators always have the full
access.
The owner, the administrators and the users with Read permission --
can read the permission of the volume.
The owner, the administrators and  the users with Write permission --
can edit the permission of the volume.
The owner and the administrators -- can give the volume access permission
to a single user, a group of users, all users.
The user can check his own access permission to a specific volume.
The permissions of the administrators are unable to edit by any other user.
The permission of the creator can be modified by administrators only.

Access permission represents as an integer, which bits are mapped to:
1st(most significant) bit -- Permission Write;
2nd -- Permission Read;
3rd -- Write;
4th -- Read.
Where

Read: the user can only read the volume, including reading the content
	of the volume, the volume properties and the volume metadata.
Write: the user can read and write the volume. The user can do any operations
	on this volume, including all the API operations.
Permission Read: the user can read the permission list for this volume and
	be able to see who can access this volume with which permission level.
Permission Write: the user can read and edit the permission list for
	this volume and be able to see and edit who can access this volume
	with which permission level.

Implementation:

Part 3. API extension
    Add os-volume-acl API extension.
    Use cases:
    * create/update permission:
        POST /v2/<project_id>/os-volume-acl
        {'body':
            'volume_acl_permission': {
                 'volume_id': <volume_id>,
                 'type': 'user', # or 'group'
                 'entity_id': <entity_id>,
                 'access_permission': <1-7>,
            }
        }
    * show permission (by permission ID):
        GET /v1/<project_id>/os-volume-acl/<permission_id>
    * list all permissions(for all the volumes for which the user has
    'Permission Read' access) or for a volume (by volume ID):
        GET /v1/<project_id>/os-volume-acl/detail
        GET /v1/<project_id>/os-volume-acl/detail?volume_id=<volume_id>
    * get access permission level for the volume ID:
        GET /v1/<project_id>/os-volume-acl/access?volume_id=<volume_id>
    * delete permission:
        DELETE /v1/<project_id>/os-volume-acl/<permission_id>

blueprint volume-acl

Change-Id: I7e04f31e32ce882241726776c79e14a151a79325